refactor: summary/post type 분리
